requirements.txt
Babel==2.3.3
Django==1.9.2
Pillow==3.2.0
argparse==1.2.1
dj-database-url==0.4.1
dj-static==0.0.6
django-bootstrap3==7.0.1
django-bootstrap3-datetimepicker==2.2.3
django-phonenumber-field==1.1.0
django-registration==2.0.4
django-registration-redux==1.4
django-toolbelt==0.0.1
gunicorn==19.4.5
phonenumberslite==7.3.0
psycopg2==2.6.1
pytz==2016.3
static3==0.6.1
wsgiref==0.1.2
